+++ src/moves.rs
@@ -51,6 +51,8 @@ impl Error for MoveError {}
 
 /// Determine if a move is valid for a given position
 ///
+/// # Errors
+///
 /// Returns `Ok(())` if the move is valid, or an error if the move is invalid
 /// (see [`MoveError`] for possible errors).
 pub fn check_move(game: &NimGame, mov: &NimAction) -> Result<(), MoveError> {
@@ -208,8 +210,9 @@ fn apply_move_(game: &mut NimGame, mov: &NimAction, unchecked: bool) -> Result<(
 /// - `game` - The game state before the move is applied
 /// - `mov` - The move to apply
 ///
-/// # Returns
+/// # Errors
 ///
+/// This function returns
 /// [`Ok`] with the unit type if the move is valid and was applied successfully,
 /// an [`Err`] with the reason why the move is invalid otherwise (see [`MoveError`])
 pub fn apply_move(game: &mut NimGame, mov: &NimAction) -> Result<(), MoveError> {
@@ -236,6 +239,10 @@ pub fn apply_move(game: &mut NimGame, mov: &NimAction) -> Result<(), MoveError> 
 /// but this is possibly subject to change.
 ///
 /// Please note that the bounds checks of the [`Vec`] indices are not disabled by this function.
+///
+/// # Errors
+///
+/// If the move is invalid. See [returns](#returns) above.
 pub unsafe fn apply_move_unchecked(game: &mut NimGame, mov: &NimAction) -> Result<(), MoveError> {
     apply_move_(game, mov, true)
 }
@@ -294,6 +301,13 @@ pub unsafe fn apply_move_unchecked(game: &mut NimGame, mov: &NimAction) -> Resul
 /// assert_eq!(moves[2].stack_index, 0);
 /// assert_eq!(moves[2].split, NimSplit::No);
 /// ```
+///
+/// # Panics    
+///
+/// If a rule with [`NimAction::Place`] is encountered which also has
+/// [`Split::Optional`] or [`Split::Always`], this function will panic.
+///
+#[must_use]
 pub fn calculate_legal_moves(
     stacks: &[Stack],
     rules: &[NimRule],